Renamed vm-disk-loss experiment to gcp-vm-disk-loss experiment
