+++ experiments/gcp/gcp-vm-disk-loss/rbac.yaml
