+++ pkg/gcp/gcp-vm-disk-loss/types/types.go
